requirements.txt
annotated-types    0.7.0    
anyio              4.8.0    
certifi            2025.1.31
charset-normalizer 3.4.1    
click              8.1.8    
colorama           0.4.6    
distro             1.9.0    
fastapi            0.115.11
greenlet           3.1.1 
h11                0.14.0
httpcore           1.0.7 
httpx              0.28.1
idna               3.10  
jiter              0.9.0 
openai             1.66.0
pip                25.0.1
pydantic           2.10.6
pydantic_core      2.27.2
PyMuPDF            1.25.3
PyPDF2             3.0.1
python-dotenv      1.0.1
requests           2.32.3
sniffio            1.3.1
SQLAlchemy         2.0.39
starlette          0.46.0
tqdm               4.67.1
typing_extensions  4.12.2
urllib3            2.3.0
uvicorn            0.34.0
fastapi~=0.115.11
pydantic~=2.10.6
SQLAlchemy~=2.0.39
openai~=1.66.0
markdown2~=2.5.3